mmConvertImage command - move functions into mmsolver::image.
