+++ src/mmSolver/image/image_convert.h
@@ -1,0 +1,39 @@
+/*
+ * Copyright (C) 2022, 2024 David Cattermole.
+ *
+ * This file is part of mmSolver.
+ *
+ * mmSolver is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * mmSolver is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with mmSolver.  If not, see <https://www.gnu.org/licenses/>.
+ * ====================================================================
+ *
+ */
+
+#ifndef MM_SOLVER_IMAGE_IMAGE_CONVERT_H
+#define MM_SOLVER_IMAGE_IMAGE_CONVERT_H
+
+// Maya
+#include <maya/MString.h>
+
+namespace mmsolver {
+namespace image {
+
+MStatus convert_image(const MString &src_file_path,
+                      const MString &dst_file_path,
+                      const MString &dst_output_format,
+                      const double resize_scale);
+
+}  // namespace image
+}  // namespace mmsolver
+
+#endif  // MM_SOLVER_IMAGE_IMAGE_CONVERT_H
